pyproject.toml
[tool.poetry]
name = "djangodrf-project"
version = "0.1.0"
description = ""
authors = ["RomanBogdanov <lackroy5111@gmail.com>"]
readme = "README.md"
packages = [{include = "djangodrf_project"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
psycopg2 = "^2.9.6"
pillow = "^10.0.0"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
django-filter = "^23.2"
djangorestframework-simplejwt = "^5.2.2"
validators = "^0.21.2"
drf-yasg = "^1.21.7"
requests = "^2.31.0"
stripe = "^5.5.0"
celery = "^5.3.1"
django-celery-beat = "^2.5.0"
redis = "^5.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
flake8-bugbear = "^23.7.10"
pep8-naming = "^0.13.3"
flake8-commas = "^2.1.0"
flake8-multiline-containers = "^0.0.19"
flake8-class-attributes-order = "^0.1.3"
flake8-clean-block = "^0.1.2"
flake8-indent-in-def = "^0.1.2"
flake8-newspaper-style = "^0.4.3"
flake8-return = "^1.2.0"
flake8-length = "^0.3.1"
isort = "^5.12.0"
flake8-quotes = "^3.3.2"
coverage = "^7.2.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
